Get remote contact information
Inside App set useEffect to run an async fetch rewuest to random.me
for contact data which is then held in state.

Map contacts onto ContactExp and ContactSm to return an array of one
expanded and n-1 non-expanded contacts to populate the display.

Set up dummy expandContact function which is passed to ContactSm
bound to the contact id via an arrow function.